chore(checkout): re organize file
