+++ src/cart/dto/checkout.dto.ts
@@ -1,6 +1,6 @@
 
 
-import { isNotEmpty, IsNotEmpty, IsNumber, IsString } from 'class-validator';
+import { IsNotEmpty,  IsString } from 'class-validator';
 
 export class CartCheckOutDTO {
     @IsNotEmpty()
